feat: bom维护 接口

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Services/Process/Bom/ProcBomService.cs
@@ -7,6 +7,7 @@
  */
 using FluentValidation;
 using Google.Protobuf.WellKnownTypes;
+using Hymson.Authentication;
 using Hymson.Infrastructure;
 using Hymson.Infrastructure.Exceptions;
 using Hymson.Infrastructure.Mapper;
@@ -17,6 +18,7 @@ using Hymson.MES.Services.Dtos.Process;
 using Hymson.Snowflake;
 using Hymson.Utils;
 using Hymson.Utils;
+using Microsoft.AspNetCore.Http;
 using System.Transactions;
 
 namespace Hymson.MES.Services.Services.Process
@@ -36,9 +38,11 @@ namespace Hymson.MES.Services.Services.Process
         private readonly IProcBomDetailRepository _procBomDetailRepository;
         private readonly IProcBomDetailReplaceMaterialRepository _procBomDetailReplaceMaterialRepository;
 
+        private readonly ICurrentUser _currentUser;
 
-        public ProcBomService(IProcBomRepository procBomRepository, AbstractValidator<ProcBomCreateDto> validationCreateRules, AbstractValidator<ProcBomModifyDto> validationModifyRules, IProcBomDetailRepository procBomDetailRepository, IProcBomDetailReplaceMaterialRepository procBomDetailReplaceMaterialRepository)
+        public ProcBomService(ICurrentUser currentUser,IProcBomRepository procBomRepository, AbstractValidator<ProcBomCreateDto> validationCreateRules, AbstractValidator<ProcBomModifyDto> validationModifyRules, IProcBomDetailRepository procBomDetailRepository, IProcBomDetailReplaceMaterialRepository procBomDetailReplaceMaterialRepository)
         {
+            _currentUser = currentUser;
             _procBomRepository = procBomRepository;
             _validationCreateRules = validationCreateRules;
             _validationModifyRules = validationModifyRules;
@@ -75,8 +79,8 @@ namespace Hymson.MES.Services.Services.Process
             //DTO转换实体
             var procBomEntity = procBomCreateDto.ToEntity<ProcBomEntity>();
             procBomEntity.Id= IdGenProvider.Instance.CreateId();
-            procBomEntity.CreatedBy = "TODO";
-            procBomEntity.UpdatedBy = "TODO";
+            procBomEntity.CreatedBy = _currentUser.UserName;
+            procBomEntity.UpdatedBy = _currentUser.UserName;
             procBomEntity.CreatedOn = HymsonClock.Now();
             procBomEntity.UpdatedOn = HymsonClock.Now();
             procBomEntity.BomCode = procBomEntity.BomCode.ToUpperInvariant();
@@ -201,15 +205,28 @@ namespace Hymson.MES.Services.Services.Process
                     }
                 }
 
-                await _procBomRepository.InsertAsync(procBomEntity);
+                response = await _procBomRepository.InsertAsync(procBomEntity);
+
+                if (response == 0)
+                {
+                    throw new BusinessException(ErrorCode.MES10602);
+                }
 
                 if (bomDetails.Count > 0)
                 {
-                    await _procBomDetailRepository.InsertsAsync(bomDetails);
+                    response = await _procBomDetailRepository.InsertsAsync(bomDetails);
+                    if (response == 0)
+                    {
+                        throw new BusinessException(ErrorCode.MES10602);
+                    }
                 }
                 if (bomReplaceDetails.Count > 0)
                 {
-                    await _procBomDetailReplaceMaterialRepository.InsertsAsync(bomReplaceDetails);
+                    response = await _procBomDetailReplaceMaterialRepository.InsertsAsync(bomReplaceDetails);
+                    if (response == 0)
+                    {
+                        throw new BusinessException(ErrorCode.MES10602);
+                    }
                 }
 
                 ts.Complete();
@@ -235,7 +252,21 @@ namespace Hymson.MES.Services.Services.Process
         /// <returns></returns>
         public async Task<int> DeletesProcBomAsync(string ids)
         {
+            var updateBy = _currentUser.UserName;
             var idsArr = StringExtension.SpitLongArrary(ids);
+
+            if (idsArr.Length < 1)
+            {
+                throw new ValidationException(ErrorCode.MES10610);
+            }
+
+            //判断需要删除的Bom是否是启用状态
+            var bomList = await _procBomRepository.GetByIdsAsync(idsArr);
+            if (bomList.Any(x => x.Status == "1")) 
+            {
+                throw new BusinessException(ErrorCode.MES10611);
+            }
+
             return await _procBomRepository.DeletesAsync(idsArr);
         }
 
@@ -300,10 +331,170 @@ namespace Hymson.MES.Services.Services.Process
 
             //DTO转换实体
             var procBomEntity = procBomModifyDto.ToEntity<ProcBomEntity>();
-            procBomEntity.UpdatedBy = "TODO";
+            procBomEntity.UpdatedBy = _currentUser.UserName;
             procBomEntity.UpdatedOn = HymsonClock.Now();
 
-            await _procBomRepository.UpdateAsync(procBomEntity);
+            //判断该bom是否还存在
+            var modelOrigin = await _procBomRepository.GetByIdAsync(procBomEntity.Id);
+            if (modelOrigin == null)
+            {
+
+                //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                //apiResult.Msg = $"此Bom在系统中不存在!";
+                //return apiResult;
+            }
+
+            //验证是否存在
+            var exists = (await _procBomRepository.GetProcBomEntitiesAsync(new ProcBomQuery()
+            {
+                SiteCode = procBomEntity.SiteCode,
+                BomCode = procBomEntity.BomCode,
+                Version = procBomEntity.Version,
+            })).Where(x=>x.Id!= procBomEntity.Id && x.IsDeleted==false) ;
+            if (exists != null && exists.Count() > 0)
+            {
+                throw new BusinessException(ErrorCode.MES10601).WithData("bomCode", procBomEntity.BomCode).WithData("version", procBomEntity.Version);
+            }
+
+            var materialList = procBomModifyDto.MaterialList.ToList();
+            var bomDetails = new List<ProcBomDetailEntity>();
+            var bomReplaceDetails = new List<ProcBomDetailReplaceMaterialEntity>();
+            if (materialList.Count > 0)
+            {
+                if (materialList.Any(a => string.IsNullOrWhiteSpace(a.MaterialCode)))
+                {
+                    throw new ValidationException(ErrorCode.MES10603);
+                }
+                if (materialList.Any(a => a.IsMain != 1 && string.IsNullOrWhiteSpace(a.ReplaceMaterialId)))
+                {
+                    throw new ValidationException(ErrorCode.MES10604);
+                    //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                    //apiResult.Msg = $"替代物料编码不能为空!";
+                    //return apiResult;
+                }
+                var mainList = materialList.Where(a => a.IsMain == 1).ToList();
+                if (mainList.Any(a => string.IsNullOrWhiteSpace(a.Code) || a.ProcedureBomId == "0"))
+                {
+                    throw new ValidationException(ErrorCode.MES10605);
+                    //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                    //apiResult.Msg = $"工序不能为空!";
+                    //return apiResult;
+                }
+                if (mainList.GroupBy(m => new { m.MaterialId, m.ProcedureBomId }).Where(g => g.Count() > 1).Count() > 0)
+                {
+                    throw new ValidationException(ErrorCode.MES10606);
+                    //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                    //apiResult.Msg = $"主物料编码+工序不能重复!";
+                    //return apiResult;
+                }
+                if (materialList.Any(a => a.MaterialId == a.ReplaceMaterialId))
+                {
+                    throw new ValidationException(ErrorCode.MES10607);
+                    //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                    //apiResult.Msg = $"替代物料不能跟主物料重复!";
+                    //return apiResult;
+                }
+
+                var replaceList = materialList.Where(a => a.IsMain == 0).ToList();
+                if (replaceList.GroupBy(m => new { m.MaterialId, m.ReplaceMaterialId }).Where(g => g.Count() > 1).Count() > 0)
+                {
+                    throw new ValidationException(ErrorCode.MES10608);
+                    //apiResult.Code = (int)ResultCode.PARAM_ERROR;
+                    //apiResult.Msg = $"主物料关联的替代物料不能重复!";
+                    //return apiResult;
+                }
+
+                long mainId = 0;
+                foreach (var material in materialList)
+                {
+                    if (material.IsMain == 1)
+                    {
+                        var bomdeail = new ProcBomDetailEntity
+                        {
+                            Id = IdGenProvider.Instance.CreateId(),
+                            SiteCode = procBomModifyDto.SiteCode,
+                            BomId = procBomEntity.Id,
+                            MaterialId = material.MaterialId.ParseToLong(),
+                            ProcedureBomId = material.ProcedureBomId.ParseToLong(),
+                            ReferencePoint = material.ReferencePoint,
+                            Usages = material.Usages,
+                            Loss = material.Loss ?? 0,
+                            CreatedBy = procBomEntity.CreatedBy,
+                            CreatedOn = procBomEntity.CreatedOn
+                        };
+                        mainId = bomdeail.Id;
+                        bomDetails.Add(bomdeail);
+                    }
+                    else
+                    {
+                        var bomReplacedeail = new ProcBomDetailReplaceMaterialEntity
+                        {
+                            Id = IdGenProvider.Instance.CreateId(),
+                            SiteCode = procBomModifyDto.SiteCode,
+                            BomId = procBomEntity.Id,
+                            BomDetailId = mainId,
+                            ReplaceMaterialId = material.ReplaceMaterialId.ParseToLong(),
+                            ReferencePoint = material.ReferencePoint,
+                            Usages = material.Usages,
+                            Loss = material.Loss ?? 0,
+                            CreatedBy = procBomEntity.CreatedBy,
+                            CreatedOn = procBomEntity.CreatedOn
+                        };
+                        bomReplaceDetails.Add(bomReplacedeail);
+                    }
+                }
+            }
+
+            #region 操作数据库
+            using (TransactionScope ts = new TransactionScope())
+            {
+                int response = 0;
+                if (procBomEntity.IsCurrentVersion)
+                {
+                    var currentProcBom = (await _procBomRepository.GetProcBomEntitiesAsync(new ProcBomQuery()
+                    {
+                        SiteCode = procBomEntity.SiteCode,
+                        BomCode = procBomEntity.BomCode,
+                    })).Where(x => x.IsCurrentVersion = true).First(); ;//a => a.SiteCode == parm.SiteCode && a.BomCode == procBom.BomCode && a.IsCurrentVersion == true
+                    if (currentProcBom != null)
+                    {
+                        await _procBomRepository.UpdateIsCurrentVersionIsFalseAsync(new long[] { currentProcBom.Id });
+                    }
+                }
+
+                //编码+版本唯一在修改时两者均不可修改
+                response = await _procBomRepository.UpdateAsync(procBomEntity);
+
+                if (response == 0) 
+                {
+                    throw new BusinessException(ErrorCode.MES10609);
+                }
+
+                await _procBomDetailRepository.DeleteBomIDAsync(new long[] { procBomEntity.Id });
+                await _procBomDetailReplaceMaterialRepository.DeleteBomIDAsync(new long[] { procBomEntity.Id });
+
+                if (bomDetails.Count > 0)
+                {
+                    response = await _procBomDetailRepository.InsertsAsync(bomDetails);
+
+                    if (response == 0)
+                    {
+                        throw new BusinessException(ErrorCode.MES10609);
+                    }
+                }
+                if (bomReplaceDetails.Count > 0)
+                {
+                    response = await _procBomDetailReplaceMaterialRepository.InsertsAsync(bomReplaceDetails);
+                    if (response == 0)
+                    {
+                        throw new BusinessException(ErrorCode.MES10609);
+                    }
+                }
+
+                ts.Complete();
+            }
+
+            #endregion
         }
 
         /// <summary>
